splitting actions
